Cleanup code, nicer LED blinks
NOTE: this version seems to work OK, but always tries to upgrade.

For now, the user app is recommended to invalidate the image in external
flash, so that the image is not re-flashed every time. This also makes
most resets quite fast: if there is no external image, there is no slow
signature verification step.